requirements.txt
# Dependencies for Age of Empires 2 Team Balancing Bot

# Discord.py for bot functionality
discord.py>=2.3.0

# HTTP client for API interactions
aiohttp>=3.8.5
requests>=2.31.0

# Database management
sqlalchemy>=2.0.20
alembic>=1.12.0

# Data processing and analysis
numpy>=1.24.0
pandas>=2.1.0

# Caching
cachetools>=5.3.1

# Configuration management
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1

# Development tools
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0 